comments and consolidations
